Update meta data.
